requirements.txt
# Core Framework
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Video Generation - CogVideoX via Diffusers
diffusers>=0.26.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
sentencepiece>=0.1.99

# Text-to-Speech - Coqui XTTS-v2
TTS>=0.21.0

# Video/Audio Processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
librosa>=0.10.0
soundfile>=0.12.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# Wav2Lip Dependencies
face-alignment>=1.3.5
batch-face>=1.4.0

# Real-ESRGAN Upscaling
realesrgan>=0.3.0
basicsr>=1.4.2

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.66.0
httpx>=0.26.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-timeout>=2.2.0
